SIGNIFICANT REDESIGN...
Now modules are responsible for connection management, determining whether
a peer is connected or not, and sharing subscription information. This allows
for greater flexibility for connectionless, connection-full, and other
network technologies.

This also collapses the sbn_netif code into sbn_app; and some of the methods
may become static in the future.